% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcov.R
\name{pcov.test}
\alias{pcov.test}
\title{Projection correlation independence test}
\usage{
pcov.test(x, y, num.permutations = 99, test.statistics = FALSE, seed = 1)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{y}{a numeric vector or matrix}

\item{num.permutations}{the number of permutation replications. 
When \code{num.permutations = 0}, the function just returns the statistic. 
Default: \code{num.permutations = 199}.}

\item{test.statistics}{a logical value. If \code{test.statistics = TRUE}, an adjusted test statistic would be computed. 
Emprically, the adjusted projection correlation can improve the statistical power but increase computational burden. 
Default: \code{test.statistics = FALSE}.}
}
\value{
If \code{num.permutations > 0}, the function returns a \code{htest} class object containing the following components:
If \code{num.permutations = 0}, the function returns a statistic value.
}
\description{
Permutation based projection correlation independence test
}
\examples{
set.seed(1)
num <- 30

## univariate case:
x <- rnorm(num)
y <- rnorm(num)
pcov.test(x, y)
y <- x + rnorm(num)
pcov.test(x, y)

## multivariate case:
p <- 3
x <- matrix(rnorm(num * p), nrow = num)
y <- matrix(rnorm(num * p), nrow = num)
pcov.test(x, y)
y <- x + matrix(rnorm(num * p), nrow = num)
pcov.test(x, y)
}
\references{
Liping Zhu, Kai Xu, Runze Li, Wei Zhong, Projection correlation between two random vectors, Biometrika, Volume 104, Issue 4, December 2017, Pages 829-843, https://doi.org/10.1093/biomet/asx043
}
